// shared/api/generated/model/getMainPropertyPropertyType.ts
/**
 * Generated by orval v6.31.0 🍺
 * Do not edit manually.
 * Dingle Swagger API
 * Dingle API
 * OpenAPI spec version: v1.0.0
 */

export type GetMainPropertyPropertyType = typeof GetMainPropertyPropertyType[keyof typeof GetMainPropertyPropertyType];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const GetMainPropertyPropertyType = {
  ONE_ROOM: 'ONE_ROOM',
  APT: 'APT',
  VILLA: 'VILLA',
  OFFICETEL: 'OFFICETEL',
} as const;