// shared/api/generated/model/dealInfoDealType.ts
/**
 * Generated by orval v6.31.0 🍺
 * Do not edit manually.
 * Dingle Swagger API
 * Dingle API
 * OpenAPI spec version: v1.0.0
 */

export type DealInfoDealType = typeof DealInfoDealType[keyof typeof DealInfoDealType];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const DealInfoDealType = {
  SALE: 'SALE',
  LEASE: 'LEASE',
  RENT: 'RENT',
} as const;